velocidad del disco segun el tiemop q haya apsado P1

+++ Game/Project_9_Solution/Project_9_Solution/Source/ModuleFrisbee.h
@@ -37,9 +37,9 @@ public:
 	iPoint position;
 
 	// The speed in which we move the player (pixels per frame)
-	int speed;
-	int xspeed;
-	int yspeed;
+	float speed;
+	float xspeed;
+	float yspeed;
 	int mov;
 	int arbitro;
